Добавлена правильная обработка изображений

+++ app/src/main/java/com/mxpj/speedyart/data/mapper/PictureMapper.kt
@@ -31,7 +31,7 @@ class PictureMapper @Inject constructor(
     ): Picture {
         return Picture(
             pictureWithDifficulties.picture.id,
-            pictureWithDifficulties.picture.assetLink.toInt(),
+            pictureWithDifficulties.picture.assetLink,
             difficultyLevelMapper.mapCompletionWithDifficultyListToDifficultyLevelList(
                 pictureWithDifficulties.completionWithDifficultyList
             )
